+++ app/Http/Controllers/Supplier/SupplierIndexController.php
@@ -14,7 +14,7 @@ class SupplierIndexController extends MainController
     public function index(Request $request)
     {
         $supplier = new Supplier();
-        $data['suppliers'] = $supplier->getAllByCompany(Auth::user()->company_id);
+        $data['suppliers'] = $supplier->getAllByCompany((int) Auth::user()->company_id);
 
         return view('supplier.index', $data);
     }
